fix: 보관함 부분 전체 width 잘리는 문제 해결
