+++ app/src/main/java/com/artem/animationjikan/presentation/ui/tab/home/Home.kt
@@ -1,19 +1,17 @@
 package com.artem.animationjikan.presentation.ui.tab.home
 
 import android.widget.Toast
+import androidx.annotation.StringRes
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
@@ -25,6 +23,7 @@ import androidx.compose.material3.SuggestionChipDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.State
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -41,13 +40,19 @@ import androidx.compose.ui.unit.sp
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.artem.animationjikan.R
+import com.artem.animationjikan.domain.entities.HomeCommonEntity
 import com.artem.animationjikan.presentation.ui.LocalNavScreenController
+import com.artem.animationjikan.presentation.ui.components.HeightGap
+import com.artem.animationjikan.presentation.ui.components.WidthGap
 import com.artem.animationjikan.presentation.ui.tab.home.components.ContentSectionRow
 import com.artem.animationjikan.presentation.ui.tab.home.components.RecommendPager
 import com.artem.animationjikan.presentation.ui.theme.AnimationJikanTheme
 import com.artem.animationjikan.util.CATEGORIES_LIST
+import com.artem.animationjikan.util.enums.ViewModelState
 import com.artem.animationjikan.util.event.UiEvent
 import com.artem.animationjikan.util.router.NavRoutes
+import com.google.gson.Gson
+import java.util.Base64
 
 @Composable
 fun HomeTab(
@@ -58,6 +63,19 @@ fun HomeTab(
     val scrollState = rememberScrollState()
     val recommendAnimationList = viewModel.recommendationAnimationList.collectAsStateWithLifecycle()
     val navController = LocalNavScreenController.current
+
+    val animeNavigationClick: (HomeCommonEntity) -> Unit = { entity ->
+        val jsonString = Gson().toJson(entity)
+        val encodedString = Base64.getUrlEncoder()
+
+        navController.navigate(
+            NavRoutes.AnimationDetail.router + "/" + encodedString.encodeToString(
+                jsonString.toByteArray(
+                    Charsets.UTF_8
+                )
+            )
+        )
+    }
 
     LaunchedEffect(key1 = Unit) {
         viewModel.eventFlow.collect { event ->
@@ -78,80 +96,62 @@ fun HomeTab(
             .padding(horizontal = 10.dp)
     ) {
         Column {
-            Spacer(modifier = Modifier.height(6.dp))
+            HeightGap(6)
 
             SearchView()
 
-            Spacer(modifier = Modifier.height(6.dp))
+            HeightGap(6)
 
             ChipSection()
 
-            Spacer(modifier = Modifier.height(16.dp))
+            HeightGap(16)
 
             RecommendPager(
                 recommendationAnimations = recommendAnimationList.value,
                 isLoading = viewModel.state != ViewModelState.Success && viewModel.recommendationAnimationList.collectAsState().value.isEmpty()
             )
 
-            Spacer(modifier = Modifier.height(25.dp))
+            HeightGap(25)
 
-            ContentSectionRow(
-                R.string.section_recently_viewed,
-                viewModel.topAnimationList.collectAsStateWithLifecycle().value,
-                isLoadingState = viewModel.state != ViewModelState.Success && viewModel.topAnimationList.collectAsStateWithLifecycle().value.isEmpty(),
-                onItemClick = { entity ->
-                    navController.navigate(NavRoutes.AnimationDetail.router + "/${entity.id}")
-                },
-                onItemLikeClick = { viewModel.toggleLike(entity = it) }
+            HomeContentSection(
+                titleRes = R.string.section_recently_viewed,
+                listState = viewModel.topAnimationList.collectAsStateWithLifecycle(),
+                onItemLikeClick = { viewModel.toggleLike(entity = it) },
+                onItemClick = animeNavigationClick,
+                viewModel = viewModel
             )
 
-            Spacer(modifier = Modifier.height(16.dp))
-
-            ContentSectionRow(
-                R.string.section_upcoming_anime,
-                viewModel.upcomingList.collectAsStateWithLifecycle().value,
-                isLoadingState = viewModel.state != ViewModelState.Success && viewModel.upcomingList.collectAsStateWithLifecycle().value.isEmpty(),
-                onItemClick = { entity ->
-                    navController.navigate(NavRoutes.AnimationDetail.router + "/${entity.id}")
-                },
-                onItemLikeClick = { viewModel.toggleLike(entity = it) }
+            HomeContentSection(
+                titleRes = R.string.section_upcoming_anime,
+                listState = viewModel.upcomingList.collectAsStateWithLifecycle(),
+                onItemLikeClick = { viewModel.toggleLike(entity = it) },
+                onItemClick = animeNavigationClick,
+                viewModel = viewModel
             )
 
-            Spacer(modifier = Modifier.height(16.dp))
-
-            ContentSectionRow(
-                R.string.section_top_anime,
-                viewModel.topAnimationList.collectAsStateWithLifecycle().value,
-                isLoadingState = viewModel.state != ViewModelState.Success && viewModel.topAnimationList.collectAsStateWithLifecycle().value.isEmpty(),
-                onItemClick = { entity ->
-                    navController.navigate(NavRoutes.AnimationDetail.router + "/${entity.id}")
-                },
-                onItemLikeClick = { viewModel.toggleLike(entity = it) }
+            HomeContentSection(
+                titleRes = R.string.section_top_anime,
+                listState = viewModel.topAnimationList.collectAsStateWithLifecycle(),
+                onItemLikeClick = { viewModel.toggleLike(entity = it) },
+                onItemClick = animeNavigationClick,
+                viewModel = viewModel
             )
 
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            ContentSectionRow(
-                R.string.section_top_manga,
-                viewModel.topMangaList.collectAsStateWithLifecycle().value,
-                isLoadingState = viewModel.state != ViewModelState.Success && viewModel.topMangaList.collectAsStateWithLifecycle().value.isEmpty(),
+            HomeContentSection(
+                titleRes = R.string.section_top_manga,
+                listState = viewModel.topMangaList.collectAsStateWithLifecycle(),
+                onItemLikeClick = { viewModel.toggleLike(entity = it) },
                 onItemClick = { _ -> },
-                onItemLikeClick = { viewModel.toggleLike(entity = it) }
+                viewModel = viewModel
             )
 
-            Spacer(modifier = Modifier.height(16.dp))
-
-            ContentSectionRow(
-                R.string.section_top_character,
-                viewModel.topCharacterList.collectAsStateWithLifecycle().value,
-                isLoadingState = viewModel.state != ViewModelState.Success && viewModel.topCharacterList.collectAsStateWithLifecycle().value.isEmpty(),
+            HomeContentSection(
+                titleRes = R.string.section_top_character,
+                listState = viewModel.topCharacterList.collectAsStateWithLifecycle(),
+                onItemLikeClick = { viewModel.toggleLike(entity = it) },
                 onItemClick = { _ -> },
-                onItemLikeClick = { viewModel.toggleLike(entity = it) }
+                viewModel = viewModel
             )
-
-
-            Spacer(modifier = Modifier.height(30.dp))
 
         }
     }
@@ -169,9 +169,7 @@ fun SearchView() {
                 width = 1.dp,
                 color = colorResource(R.color.grey4),
                 shape = RoundedCornerShape(8.dp)
-            ).clickable {
-                //TODO navigate to search screen
-            }
+            )
 
     ) {
         Row(
@@ -188,9 +186,9 @@ fun SearchView() {
                 modifier = Modifier.size(18.dp),
                 contentDescription = stringResource(R.string.search)
             )
-            Spacer(modifier = Modifier.width(8.dp))
+            WidthGap(8)
             Text(
-                "Search Contents",
+                stringResource(R.string.search_default_txt),
                 modifier = Modifier.weight(1f),
                 fontSize = 11.sp,
                 fontWeight = FontWeight(400),
@@ -219,6 +217,27 @@ fun ChipSection() {
             )
         }
     }
+}
+
+@Composable
+private fun HomeContentSection(
+    @StringRes titleRes: Int, // 섹션 제목만 다름
+    listState: State<List<HomeCommonEntity>>,
+    onItemClick: (HomeCommonEntity) -> Unit,
+    onItemLikeClick: (HomeCommonEntity) -> Unit,
+    viewModel: HomeTabViewModel
+) {
+    val listValue = listState.value
+    val isLoading = viewModel.state != ViewModelState.Success && listValue.isEmpty()
+
+    ContentSectionRow(
+        title = titleRes,
+        list = listValue,
+        isLoadingState = isLoading,
+        onItemClick = onItemClick,
+        onItemLikeClick = onItemLikeClick
+    )
+    HeightGap(16)
 }
